unreliablefs: allow arbitrary large duration values
`nanosleep()` fails if `.tv_nsec` is larger than 999999999, so duration
must be distributed between `.tv_sec` and `.tv_nsec`. Duration also
needs to be stored in `unsigned long long` to be able to store values of
more than ~4 seconds.

+++ unreliablefs/unreliablefs_errinj.h
@@ -54,7 +54,7 @@ struct errinj_conf {
     char *path_regexp;
     char *errno_regexp;
     unsigned int probability;
-    unsigned int duration;
+    unsigned long long duration;
     errinj_type type;
 
     TAILQ_ENTRY(errinj_conf) entries;
